backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
cryptography
pydantic
supabase
python-dotenv
bcrypt
